requirements copy.txt
Django
djangorestframework
gunicorn
dj-database-url
psycopg2-binary
dj-database-url==0.5.0
